Recursion optimization v1
