+++ tempit/core.py
@@ -56,11 +56,16 @@ def tempit(
         verbose: bool = False,
         check_for_recursion: bool = False,
     ) -> Callable:
+        potential_recursion_func_stack: List[Callable] = []
+
         @wraps(func)
         def tempit_wrapper(*args: Tuple, **kwargs: Dict) -> Any:
-            is_recursive, run_times_final, concurrent_execution_final = check_is_recursive_func(
-                check_for_recursion, func, run_times, concurrent_execution
+            nonlocal potential_recursion_func_stack
+
+            potential_recursion_func_stack, run_times_final, concurrent_execution_final = check_is_recursive_func(
+                check_for_recursion, func, run_times, concurrent_execution, potential_recursion_func_stack
             )
+
             callable_func, args, args_to_print = extract_callable_and_args_if_method(func, *args)
             result, total_times, real_time = function_execution(
                 callable_func,
@@ -69,10 +74,16 @@ def tempit(
                 *args,
                 **kwargs,
             )
-            if not is_recursive:
+
+            if potential_recursion_func_stack:
+                # Remove it from the potential recursion stack
+                potential_recursion_func_stack.pop()
+
+            if not potential_recursion_func_stack:
                 print_tempit_values(
                     run_times_final, verbose, callable_func, total_times, real_time, *args_to_print, **kwargs
                 )
+
             return result
 
         return tempit_wrapper
@@ -97,27 +108,41 @@ def tempit(
 
 
 def check_is_recursive_func(
-    check_for_recursion: bool, func: Callable, run_times: int, concurrent_execution: bool
-) -> Tuple[bool, int, bool]:
+    check_for_recursion: bool,
+    func: Callable,
+    run_times: int,
+    concurrent_execution: bool,
+    potential_recursion_func_stack: List[Callable],
+) -> Tuple[List[Callable], int, bool]:
     """
     Checks if the function is being called recursively.
     Returns:
-        Tuple[bool, int, bool]: A tuple containing True if the function is being called recursively, False otherwise.
-        The second element is the number of times the function has been called, and the third element is a boolean indicating if the function has crashed.
+        Tuple[List[Callable], int, bool]: A tuple containing the potential recursive function stack to check if the function is being called recursively, None otherwise.
+        The second element is the run_times parameter, and the third element is a boolean indicating if concurrent_execution is enabled.
     """
 
+    if potential_recursion_func_stack:
+        if potential_recursion_func_stack[-1] == func:
+            # Check if the function is being called recursively by checking the object identity. This is way faster than useing getFrameInfo
+            potential_recursion_func_stack.append(func)
+            return potential_recursion_func_stack, 1, False
+    else:
+        potential_recursion_func_stack.append(func)
+    """
     if check_for_recursion:
         import sys
         from inspect import getframeinfo
 
+        print("HERE")
         func_name = func.__name__
         func_filename = ""
         if hasattr(func, "__code__"):
             func_filename = func.__code__.co_filename
         frame = getframeinfo(sys._getframe(2), context=0)
         if frame.function == func_name and func_filename == frame.filename:
-            return True, 1, False
-    return False, run_times, concurrent_execution
+            return potential_recursion_func_stack, 1, False
+    """
+    return potential_recursion_func_stack, run_times, concurrent_execution
 
 
 def extract_callable_and_args_if_method(func, *args) -> Tuple[Callable, Tuple, Tuple]:
